% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_filepath}
\alias{check_filepath}
\title{Checks If Filepath is Valid, Optionally Fixes It.}
\usage{
check_filepath(filepath = filepath)
}
\arguments{
\item{filepath}{Character vector of filepaths to check.}
}
\value{
filepath as a named character vector
}
\description{
Checks If Filepath is Valid, Optionally Fixes It.
}
