% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSX-package.R
\docType{package}
\name{readSX-package}
\alias{readSX}
\alias{readSX-package}
\title{readSX: Read 'SurveyXact' Data}
\description{
Imports data from the 'SurveyXact' commercial service \url{https://rambollxact.com}, adds variable labels, and converts value labels to factors in the same order as exported from 'SurveyXact'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIFU-NO/readSX}
  \item \url{https://NIFU-NO.github.io/readSX/}
  \item Report bugs at \url{https://github.com/NIFU-NO/readSX/issues}
}

}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID})

}
\keyword{internal}
