% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_for_sx.R
\name{write_data_for_sx}
\alias{write_data_for_sx}
\title{Writes Data with Respondents in CSV Format Required for SurveyXact Import}
\usage{
write_data_for_sx(data, filepath, overwrite = FALSE)
}
\arguments{
\item{data}{Data frame (or tibble).}

\item{filepath}{Filepath as string}

\item{overwrite}{Whether to overwrite existing file or not (default).}
}
\value{
Side-effect: writes file. Returns filepath.
}
\description{
Writes Data with Respondents in CSV Format Required for SurveyXact Import
}
\examples{
tmpfile <- tempfile(fileext = ".csv")
write_data_for_sx(mtcars, filepath=tmpfile)
unlink(tmpfile)
}
