% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSX-package.R
\docType{package}
\name{readSX-package}
\alias{readSX}
\alias{readSX-package}
\title{readSX: Read SurveyXact Data into R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID})

}
\keyword{internal}
