% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSX-package.R
\docType{package}
\name{readSX-package}
\alias{readSX}
\alias{readSX-package}
\title{readSX: Read SurveyXact Data into R}
\description{
Imports data from the SurveyXact service, dealing with labels correctly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sda030/readSX}
  \item Report bugs at \url{https://github.com/sda030/readSX/issues}
}

}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID})

}
\keyword{internal}
