% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSX-package.R
\docType{package}
\name{readSX-package}
\alias{readSX}
\alias{readSX-package}
\title{readSX: Read 'SurveyXact' Data}
\description{
Imports data from the 'SurveyXact' commercial service\url{https://www.surveyxact.com/}. It adds variable labels and value labels, which should be converted to e.g. factors with labelled::to_factor().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sda030/readSX}
  \item Report bugs at \url{https://github.com/sda030/readSX/issues}
}

}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID})

}
\keyword{internal}
